refactor: ♻️ refactoring card collections + ui touchups
