+++ pages/[user]/components/collections/CollectionFolder.tsx
@@ -5,17 +5,17 @@ import {
   Group,
   Stack,
   Text,
-  ThemeIcon,
   Tooltip,
   useMantineTheme,
 } from "@mantine/core";
 import Link from "next/link";
-import { IconUsers, IconHeart, IconEyeOff, IconEye } from "@tabler/icons";
 
 import CollectionCard from "./CollectionCard";
 import { ICONS } from "./icons";
+import PRIVACY_LEVELS from "./privacyLevels";
 
 import { CardCollection, CollectedCard } from "types/makotools";
+import ResponsiveGrid from "components/core/ResponsiveGrid";
 
 function CollectionFolder({
   collection,
@@ -31,7 +31,7 @@ function CollectionFolder({
   };
 
   const icon = ICONS[collection.icon || 0];
-
+  const privacy = PRIVACY_LEVELS[collection.privacyLevel];
   return (
     <Accordion.Item value={collection.id}>
       <Accordion.Control py="xs" px="md">
@@ -42,60 +42,43 @@ function CollectionFolder({
           <Box>
             <Text size="md" weight={800}>
               {collection.name}
+              {isYourProfile && (
+                <Tooltip
+                  label={
+                    <Text size="xs" weight={500}>
+                      {privacy.description}
+                    </Text>
+                  }
+                  position="top"
+                  withinPortal
+                >
+                  <Text
+                    inline
+                    span
+                    ml="xs"
+                    sx={{ verticalAlign: -2 }}
+                    color="dimmed"
+                  >
+                    <privacy.icon size={16} />
+                  </Text>
+                </Tooltip>
+              )}
             </Text>
             <Text size="xs" weight={500} color="dimmed">
               {collection.cards.length} cards in collection
             </Text>
           </Box>
-          {isYourProfile &&
-            (collection.privacyLevel === 1 ? (
-              <Tooltip
-                label="Only logged in users can view this collection."
-                withArrow
-              >
-                <ThemeIcon variant="light" color="lime" sx={ICON_STYLE}>
-                  <IconUsers size={20} />
-                </ThemeIcon>
-              </Tooltip>
-            ) : collection.privacyLevel === 2 ? (
-              <Tooltip
-                label="Only your friends can view this collection."
-                withArrow
-              >
-                <ThemeIcon variant="light" color="pink" sx={ICON_STYLE}>
-                  <IconHeart size={20} />
-                </ThemeIcon>
-              </Tooltip>
-            ) : collection.privacyLevel === 3 ? (
-              <Tooltip label="This collection is completely private." withArrow>
-                <ThemeIcon variant="light" color="gray" sx={ICON_STYLE}>
-                  <IconEyeOff size={20} />
-                </ThemeIcon>
-              </Tooltip>
-            ) : (
-              <Tooltip label="Anyone can view this collection." withArrow>
-                <ThemeIcon sx={ICON_STYLE}>
-                  <IconEye size={20} />
-                </ThemeIcon>
-              </Tooltip>
-            ))}
         </Group>
       </Accordion.Control>
       <Accordion.Panel>
         {collection.cards && collection.cards.length > 0 ? (
-          <Box
-            sx={(theme) => ({
-              display: "grid",
-              gridTemplateColumns: "repeat(auto-fill, minmax(100px, 1fr))",
-              gap: theme.spacing.xs,
-            })}
-          >
+          <ResponsiveGrid width={100}>
             {collection.cards
               .filter((c: CollectedCard) => c.count)
               .map((c: CollectedCard) => (
-                <CollectionCard key={c.id} card={c} editing={false} />
+                <CollectionCard key={c.id} card={c} />
               ))}
-          </Box>
+          </ResponsiveGrid>
         ) : (
           <Stack align="flex-start">
             <Text color="dimmed">This collection is empty.</Text>
